modify ASControlTarget
